+++ lib/src/models/issue.dart
@@ -3,6 +3,9 @@ import 'package:github_snitch/src/models/comment.dart';
 const String issueIdField = "number";
 const String issueTitleField = "title";
 const String issueStateField = "state";
+const String issueBodyField = "body";
+const String issueLabelsField = "labels";
+const String issueFieldMilstone = "milestone";
 const String issueCommentsField = "comments";
 
 class Issue {
@@ -12,6 +15,9 @@ class Issue {
   bool isOpen = false;
   Comment comments = Comment();
   List<Issue> multi = [];
+  List<String> labels = [];
+  String? body;
+  int? milestone;
 
   Issue();
 
@@ -19,6 +25,10 @@ class Issue {
     id = json[issueIdField].toString();
     title = json[issueTitleField];
     state = json[issueStateField];
+    labels =
+        List<String>.from(json[issueLabelsField]?.map((e) => e['name']) ?? []);
+    body = json[issueBodyField];
+    milestone = json[issueFieldMilstone]?['number'] ?? 0;
     isOpen = state == "open";
   }
 }