+++ CHANGELOG.md
@@ -14,3 +14,34 @@
 * Fixed show all stack issue
 * Support include screenshots on report
 * Used fine-grained personal access token instead of PAT
+
+## 0.0.5
+* Removed createLabels method
+* Add fetch issues from GitHub
+* Added string similarity package for check issue already reported
+## 0.0.6
+
+* Fix removeLastLine extension issue
+* Add string similarity package
+* Update client info package
+
+## 0.0.7
+* Added labels arg to listenToExceptions method
+
+## 0.0.8
+* Added documentations
+
+## 0.0.9
+* Use connectivity package for check network
+* Update readme
+
+## 0.0.10
+* Use connectivity package for check network
+* Configured saved issues report
+* Use device info plus package for support all platform
+
+## 0.0.11
+* Fixed bug on issue model
+
+## 0.0.12
+* Added report issue on app version milestone